[BF] 2016-11-21 晚

+++ app/src/main/java/com/soubu/goldensteward/GoldenStewardApplication.java
@@ -110,7 +110,7 @@ public class GoldenStewardApplication extends Application implements Application
         mPhone = name;
         SharedPreferences sp = AppUtil.getDefaultSharedPreference(sInstance);
         sp.edit().putString(Constant.SP_KEY_USER_PHONE, name).commit();
-        if (user != null) {
+        if (dao != null && user != null) {
             user.setPhone(name);
             dao.update(user);
         }
@@ -170,7 +170,7 @@ public class GoldenStewardApplication extends Application implements Application
     }
 
     public void clearUser() {
-        if (dao != null) {
+        if (dao != null && user != null) {
             dao.delete(user);
         }
     }
